requirements.txt
pydantic>=2.0.0
numpy>=1.24.0
tweepy>=4.14.0
tenacity>=8.2.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
